- UPDAYE OPTION NO SMS OPTION - Bug Fixed
- CANCELATION SMS IS NOT WORKING - Bug Fixed

- COMPLETE SCREEN MAKE DEFAULT "PAYMENT: PAID", "PAYMENT TYPE: CASH",
CURSOR TO GO DIRECTLY 

TO INVOICE NUMBER. - Changes Done

- ADD CUSTOMER, LAST NAME IS OPTIONAL BUT EDIT CUSTOMER LAST NAME IS
MANDATORY, MAKE THE 

EDIT SCREENT FOR LAST SCREEN OPTIONAL - Bug Fixed

- UPDATE CUSTOMER IS NOT WORKING TO CHANGE THE MOBILE NUMBER, ALSO NO
POP UP SCREEN 

- Bug Fixed

- EDIT STAFF IS NOT WORKING - It's working, no changes done. Please
check again

- PDF REPORT IS NOT WORKING - Bug Fixed

- REPORT RESET OPTION IS NOT WORKING - Bug Fixed

- REPORT ADDRESS SHOULD START AS " AREA- BULDING NAME- FLAT NO" -
Changes Done

- ONE MORE REPORT REQUIRED "MONTHLY WORK" WHICH WILL DIPLAY 'CUST NAME',
'DATE', 'STAFF', 

'AMOUNT', 'STATUS (COMPL, CANCL, CREAT)  - New Report added

+++ jbm/src/main/java/com/innovazions/jbm/entity/Appointment.java
@@ -27,7 +27,7 @@ public class Appointment extends CoreEntity<Appointment, AppointmentView>
 
 	private String cancellationReason;
 
-	private Integer hoursSpent;
+	private Float hoursSpent;
 
 	private double payableAmount;
 
@@ -90,11 +90,11 @@ public class Appointment extends CoreEntity<Appointment, AppointmentView>
 		this.cancellationReason = cancellationReason;
 	}
 
-	public Integer getHoursSpent() {
+	public Float getHoursSpent() {
 		return hoursSpent;
 	}
 
-	public void setHoursSpent(Integer hoursSpent) {
+	public void setHoursSpent(Float hoursSpent) {
 		this.hoursSpent = hoursSpent;
 	}
